tools: GXMT0005 stream format definition
Expand the object ID field inside the GXMT stream format to be
64-bit. It remains unclear how it came to be that the parent_fid
field was already 64-bit to begin with.

+++ tools/genimport.hpp
@@ -78,7 +78,7 @@ struct tgt_folder {
 };
 
 using attachment_content_ptr = std::unique_ptr<ATTACHMENT_CONTENT, gi_delete>;
-using gi_folder_map_t = std::unordered_map<uint32_t, tgt_folder>;
+using gi_folder_map_t = std::unordered_map<uint64_t, tgt_folder>;
 using message_content_ptr = std::unique_ptr<MESSAGE_CONTENT, gromox::mc_delete>;
 using propname_array_ptr = std::unique_ptr<PROPNAME_ARRAY, gi_delete>;
 using tarray_set_ptr = std::unique_ptr<TARRAY_SET, gi_delete>;
